# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api": {
      "get": {
        "operationId": "AppController_getData",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FooDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/2": {
      "get": {
        "operationId": "AppController_getDataTwo",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FooDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/booking/{type}": {
      "get": {
        "operationId": "BookingController_getAllReports",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingResponseDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "BookingController_createReport",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponseDto"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/booking/{type}/{id}": {
      "get": {
        "operationId": "BookingController_getReportById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponseDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "BookingController_updateReport",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponseDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "BookingController_deleteReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "204": {
            "description": ""
          }
        }
      }
    },
    "/api/services/{type}/foo": {
      "get": {
        "operationId": "ServicesController_index",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServiceResponseDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/services/{type}/foo{id}": {
      "get": {
        "operationId": "ServicesController_show",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/services/{type}": {
      "get": {
        "operationId": "ServicesController_getAllReports",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServiceResponseDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "ServicesController_createReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseDto"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/services/{type}/{id}": {
      "get": {
        "operationId": "ServicesController_getReportById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "ServicesController_updateReport",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponseDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "ServicesController_deleteReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "204": {
            "description": ""
          }
        }
      }
    }
  },
  "info": {
    "title": "Nx Psych Api",
    "description": "The API for Nx Psych App",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "FooDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "BookingResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "timeSlotId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "isHost": {
            "type": "boolean"
          },
          "paymentDetailsId": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "type": {
            "enum": [
              "remote",
              "inperson"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "timeSlotId",
          "userId",
          "isHost",
          "paymentDetailsId",
          "created_at",
          "updated_at",
          "type"
        ]
      },
      "CreateBookingDto": {
        "type": "object",
        "properties": {
          "timeSlotId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "isHost": {
            "type": "boolean"
          },
          "paymentDetailsId": {
            "type": "string"
          }
        },
        "required": [
          "timeSlotId",
          "name",
          "userId",
          "isHost",
          "paymentDetailsId"
        ]
      },
      "UpdateBookingDto": {
        "type": "object",
        "properties": {
          "timeSlotId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "isHost": {
            "type": "boolean"
          },
          "paymentDetailsId": {
            "type": "string"
          }
        },
        "required": [
          "timeSlotId",
          "name",
          "userId",
          "isHost",
          "paymentDetailsId"
        ]
      },
      "ServiceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "dateTime": {
            "format": "date-time",
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "venue": {
            "type": "string"
          },
          "duration": {
            "type": "number"
          },
          "price": {
            "type": "number"
          },
          "location": {
            "type": "string"
          },
          "isCancelled": {
            "type": "boolean"
          },
          "type": {
            "enum": [
              "remote",
              "inperson"
            ],
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "category",
          "type"
        ]
      },
      "CreateServiceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "dateTime": {
            "format": "date-time",
            "type": "string"
          },
          "duration": {
            "type": "number",
            "minimum": 1
          },
          "description": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "minimum": 1
          },
          "location": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "remote",
              "inperson"
            ]
          },
          "isCancelled": {
            "type": "boolean"
          },
          "city": {
            "type": "string"
          },
          "venue": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "dateTime",
          "duration",
          "description",
          "image",
          "category",
          "price",
          "location",
          "type",
          "isCancelled"
        ]
      },
      "UpdateServiceDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "dateTime": {
            "format": "date-time",
            "type": "string"
          },
          "duration": {
            "type": "number",
            "minimum": 1
          },
          "description": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "minimum": 1
          },
          "location": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "remote",
              "inperson"
            ]
          }
        },
        "required": [
          "title",
          "dateTime",
          "duration",
          "description",
          "image",
          "category",
          "price",
          "location",
          "type"
        ]
      }
    }
  }
}